wip: adjust gateway infrastructure
