+++ .gitignore
@@ -33,3 +33,5 @@ build/
 
 ### VS Code ###
 .vscode/
+node_modules
+.env
