--- backend/gateway-service/.gitignore
@@ -1,2 +1,0 @@
-node_modules
-.env
