+++ gateway/Dockerfile
@@ -1,0 +1,11 @@
+FROM node:lts-bullseye
+
+WORKDIR /app
+
+COPY package.json package-lock.json* ./
+RUN npm ci
+
+COPY . .
+
+EXPOSE 3000
+CMD ["npm", "start"]
